Updated Launcher to use passed params for API call. Migrated from Shared -> Common .netstandard2.0

+++ WorldServer/World/Abilities/Buffs/BuffInterface.cs
@@ -905,7 +905,7 @@ namespace WorldServer
             if (buffList.Count == 0)
                 return;
 
-            int curTime = TCPManager.GetTimeStamp();
+            uint curTime = (uint) TCPManager.GetTimeStamp();
 
             for (int i=0; i < buffList.Count; ++i)
             {
